DATAREDIS-425 - Move default ReferenceResolver to RedisKeyValueAdapter.
Change visibility and add as default in setup.
